split entrypoint for power bootloader
